Converted SettingsManager to Scala.
Had most of it right, but turns out I was using the singleton object
concept incorrectly and so SettingsManager type couldn’t be found. Got
help from Nathan and Marcus to fix it and then got it working. Created
a class inside the object to be able to use json calls correctly since
it gets settings and writes them to disk.