Ported ImageViewerActivity
Ported this with only a few issues. Main issue was at the end where the
chapter would not show up when clicked on. It required a manual turn
on.off on the screen to show it. click to next page, nothing happens,
needs manual off.on. Finally fixed it by looking at scaloid tutorial
and adding a run on ui thread in the future call. ALSO, changed the
timeout time to 60 seconds.