Action Bar Color!
Updated the target build version to 21. This allows us to use
colorPrimary to set the color of the action bar and status bar for
devices running Android 5.0 or higher. Set the color to the yellow hue
used by Marcus in his Celty helmet drawing. Unfortunately, white text
looks a bit odd in the yellow background. I could not get the text to
change to the color black even after a variety of methods over 2 hours
of attempts.
Also, could not get the width for the next/previous buttons to change
either.