Less Buttons!
Got rid of clear buttons in each individual pane so that all clear
buttons are in settings only. Also fixed indentation in code in places.
However, for some reason the names of the mangas are centered in the
list view. No idea why this is happening and will try to fix it.